+++ io.github.adr.embedded/src/main/java/io/github/adr/embedded/ADRs.java
@@ -1,23 +1,10 @@
-/*
- * Copyright (c) 2017-2021 Oliver Kopp, Olaf Zimmermann
- *
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v2.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v20.html
- *
- * Contributors:
- *   Oliver Kopp, Olaf Zimmermann - initial API and implementation
- */
 package io.github.adr.embedded;
 
 import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
 
-@Retention(RetentionPolicy.RUNTIME)
+@Retention(RetentionPolicy.SOURCE)
 @Documented
 public @interface ADRs {
     ADR[] value();