+++ io.github.adr.embedded/src/main/java/io/github/adr/embedded/MADR.java
@@ -1,14 +1,3 @@
-/*
- * Copyright (c) 2017 Oliver Kopp, Olaf Zimmermann
- *
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v2.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v20.html
- *
- * Contributors:
- *   Oliver Kopp, Olaf Zimmermann - initial API and implementation
- */
 package io.github.adr.embedded;
 
 import java.lang.annotation.Documented;
@@ -30,14 +19,14 @@ import java.lang.annotation.Target;
     ElementType.TYPE,
     ElementType.TYPE_PARAMETER,
     ElementType.TYPE_USE})
-// @ADR(2) // commented out, does not compile (Java 1.8/Eclipse JEE Photon)
-@Retention(RetentionPolicy.RUNTIME)
+@ADR(2)
+@Retention(RetentionPolicy.SOURCE)
 @Documented
 @Repeatable(MADRs.class)
 public @interface MADR {
 
     /**
-     * The number part of the of the ADR. For instance, 1 gets expanded to the full id <code>ADR-0001</code>.
+     * The number part of the ADR. For instance, 1 gets expanded to the full id <code>ADR-0001</code>.
      *
      * @return The number of the ADR
      */