+++ io.github.adr.embedded/src/main/java/io/github/adr/embedded/MADRs.java
@@ -1,14 +1,3 @@
-/*
- * Copyright (c) 2017 Oliver Kopp, Olaf Zimmermann
- *
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v2.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v20.html
- *
- * Contributors:
- *   Oliver Kopp, Olaf Zimmermann - initial API and implementation
- */
 package io.github.adr.embedded;
 
 import java.lang.annotation.Documented;
@@ -28,7 +17,7 @@ import java.lang.annotation.Target;
     ElementType.TYPE,
     ElementType.TYPE_PARAMETER,
     ElementType.TYPE_USE})
-@Retention(RetentionPolicy.RUNTIME)
+@Retention(RetentionPolicy.SOURCE)
 @Documented
 public @interface MADRs {